Move setting of HostInfo from IPostConfigureOptions<T> to IConfigureOptions<T>
