+++ src/Tingle.EventBus/DependencyInjection/EventBusPostConfigureOptions.cs
@@ -4,7 +4,6 @@ using System;
 using System.Linq;
 using Tingle.EventBus;
 using Tingle.EventBus.Registrations;
-using Tingle.EventBus.Serialization;
 
 namespace Microsoft.Extensions.DependencyInjection
 {
@@ -23,7 +22,7 @@ namespace Microsoft.Extensions.DependencyInjection
         /// <inheritdoc/>
         public void PostConfigure(string name, EventBusOptions options)
         {
-            // check bounds for startup delay, if provided
+            // Check bounds for startup delay, if provided
             if (options.StartupDelay != null)
             {
                 var ticks = options.StartupDelay.Value.Ticks;
@@ -32,7 +31,7 @@ namespace Microsoft.Extensions.DependencyInjection
                 options.StartupDelay = TimeSpan.FromTicks(ticks);
             }
 
-            // check bounds for duplicate detection duration, if duplicate detection is enabled
+            // Check bounds for duplicate detection duration, if duplicate detection is enabled
             if (options.EnableDeduplication)
             {
                 var ticks = options.DuplicateDetectionDuration.Ticks;
@@ -41,28 +40,19 @@ namespace Microsoft.Extensions.DependencyInjection
                 options.DuplicateDetectionDuration = TimeSpan.FromTicks(ticks);
             }
 
-            // setup HostInfo
+            // Ensure we have HostInfo set
             if (options.HostInfo == null)
             {
-                var entry = System.Reflection.Assembly.GetEntryAssembly() ?? System.Reflection.Assembly.GetCallingAssembly();
-                options.HostInfo = new HostInfo
-                {
-                    ApplicationName = environment.ApplicationName,
-                    ApplicationVersion = entry.GetName().Version.ToString(),
-                    EnvironmentName = environment.EnvironmentName,
-                    LibraryVersion = typeof(EventBus).Assembly.GetName().Version.ToString(),
-                    MachineName = Environment.MachineName,
-                    OperatingSystem = Environment.OSVersion.ToString(),
-                };
+                throw new InvalidOperationException($"'{nameof(options.HostInfo)}' must be set.");
             }
 
-            // ensure there is at least one registered transport
+            // Ensure there is at least one registered transport
             if (options.RegisteredTransportNames.Count == 0)
             {
-                throw new InvalidOperationException("There must be at least one registered transport");
+                throw new InvalidOperationException("There must be at least one registered transport.");
             }
 
-            // if the default transport name has been set, ensure it is registered
+            // If the default transport name has been set, ensure it is registered
             if (!string.IsNullOrWhiteSpace(options.DefaultTransportName))
             {
                 // ensure the transport name set has been registered
@@ -73,7 +63,7 @@ namespace Microsoft.Extensions.DependencyInjection
                 }
             }
 
-            // if the default transport name has not been set, and there is only one registered, set it as default
+            // If the default transport name has not been set, and there is only one registered, set it as default
             if (string.IsNullOrWhiteSpace(options.DefaultTransportName))
             {
                 if (options.RegisteredTransportNames.Count == 1)
@@ -82,14 +72,14 @@ namespace Microsoft.Extensions.DependencyInjection
                 }
             }
 
-            // if the ConsumerNameSource is Prefix or requires the prefix, ensure it is set
+            // If the ConsumerNameSource is Prefix or requires the prefix, ensure it is set
             if ((options.ConsumerNameSource == ConsumerNameSource.Prefix || options.ConsumerNameSource == ConsumerNameSource.PrefixAndTypeName)
                 && string.IsNullOrWhiteSpace(options.ConsumerNamePrefix))
             {
-                throw new InvalidOperationException($"'{options.ConsumerNamePrefix}' when using {nameof(ConsumerNameSource)}.{options.ConsumerNameSource}");
+                throw new InvalidOperationException($"'{nameof(options.ConsumerNamePrefix)}' when using {nameof(ConsumerNameSource)}.{options.ConsumerNameSource}");
             }
 
-            // setup names and transports for each event and its consumers
+            // Setup names and transports for each event and its consumers
             var registrations = options.Registrations.Values.ToList();
             foreach (var evr in registrations)
             {
@@ -99,7 +89,7 @@ namespace Microsoft.Extensions.DependencyInjection
                    .SetConsumerNames(options, environment); // set the consumer names
             }
 
-            // ensure there are no events with the same name
+            // Ensure there are no events with the same name
             var conflicted = registrations.GroupBy(r => r.EventName).FirstOrDefault(kvp => kvp.Count() > 1);
             if (conflicted != null)
             {
@@ -108,9 +98,9 @@ namespace Microsoft.Extensions.DependencyInjection
                                                   + $" Types:\r\n- {string.Join("\r\n- ", names)}");
             }
 
+            // Ensure there are no consumers with the same name per event
             foreach (var evr in registrations)
             {
-                // ensure there are no consumers with the same name per event
                 var conflict = evr.Consumers.GroupBy(ecr => ecr.ConsumerName).FirstOrDefault(kvp => kvp.Count() > 1);
                 if (conflict != null)
                 {
